chore: Move Candidate to anchor-service.ts
